movie_pipeline/requirements.txt
psycopg2_binary==2.9.9
uuid==1.30